Update README.md to reflect HR interface implementation for criteria management
* **HRInterface.js**
  - Create `HRInterface` component
  - Implement functionality to add, remove, and adjust assessment criteria
  - Implement functionality to define weights for each criterion
  - Add state management for criteria and weights using React hooks

* **useCriterion.js**
  - Create `useCriterion` custom React hook
  - Implement state management for criteria and newCriterion using `useState`
  - Implement `addCriterion` function to add a new criterion
  - Implement `removeCriterion` function to remove a criterion by index
  - Export `useCriterion` hook with criteria, newCriterion, setNewCriterion, addCriterion, and removeCriterion
